added service to post templates, get topicId, and store it, added alertifyjs to chapters component

+++ src/app/main/categories/chapters/chapters.component.ts
@@ -3,6 +3,8 @@ import { ParamMap, ActivatedRoute } from '@angular/router';
 import { CommonHttpService } from '../../shared/commonHttp.service';
 import { CategoriesService } from '../categories.service';
 import { ChaptersService } from './chapters.service';
+import { TemplatesService } from './templates/templates.service';
+import alertify from 'alertifyjs';
 
 @Component({
   selector: 'app-chapters',
@@ -17,48 +19,67 @@ export class ChaptersComponent implements OnInit {
   constructor(
     private route: ActivatedRoute,
     private categoriesService: CategoriesService,
-    private chaptersService: ChaptersService
+    private chaptersService: ChaptersService,
+    private templatesService: TemplatesService
 
   ) { }
 
   ngOnInit() {
 
     this.route.data
-    .subscribe((res: {cats: any}) => {
-      this.course = res.cats;
-      this.chapters = this.course.chapters;  
-    })
+      .subscribe((res: { cats: any }) => {
+        this.course = res.cats;
+        this.chapters = this.course.chapters;
+      })
 
     this.role = localStorage.getItem('role');
   }
-  
+
   deleteChapter(chapterId, i): any {
 
-    this.chaptersService.deleteChapter(chapterId)
-      .subscribe((res: any) => {
+    alertify.confirm("Do you wish to delete this chapter",
+      () => {
+        this.chaptersService.deleteChapter(chapterId)
+          .subscribe((res: any) => {
 
-        console.log(res);
-        let obj = this.chapters.splice(i, 1);
-        console.log(obj);
-      }, (err: any) => {
+            console.log(res);
+            let obj = this.chapters.splice(i, 1);
+            alertify.success('Chapter Deleted');
+            console.log(obj);
+          }, (err: any) => {
+            alertify.alert(err.msg).setHeader('Message');
+            console.log(err);
+          })
+      },
+      () => {
+        alertify.error('Cancel');
+      }).setHeader('Confirmation');
 
-        console.log(err);
-      })
+
   }
 
-// not done yet complete later
   deleteTopic(topicId, i, j): any {
 
-    this.chaptersService.deleteTopic(topicId)
-      .subscribe((res: any) => {
+    alertify.confirm("Do you wish to delete this topic",
+      () => {
+        this.chaptersService.deleteTopic(topicId)
+          .subscribe((res: any) => {
 
-        console.log(res);
-        let obj = this.chapters[i].topics.splice(j, 1);
-        console.log(obj);
-      }, (err: any) => {
+            console.log(res);
+            let obj = this.chapters[i].topics.splice(j, 1);
+            alertify.success('Topic Deleted');
+          }, (err: any) => {
+            alertify.alert(err.msg).setHeader('Message');
+            console.log(err);
+          })
+      },
+      () => {
+        alertify.error('Cancel');
+      }).setHeader('Confirmation');
+  }
 
-        console.log(err);
-      })
+  getTopicData(topicId: number): any {
+    this.chaptersService.storeTopicId(topicId);
   }
 }
